agents/vision/requirements.txt
google-genai>=0.8.0
opencv-python-headless>=4.8.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9
numpy>=1.24.0
